fix MorphirIRVersion encoder

+++ morphir/toolkit/codec/zio/json/src/org/finos/morphir/ir/json/MorphirJsonEncodingSupport.scala
@@ -537,7 +537,8 @@ trait MorphirJsonEncodingSupport {
       }
     }
 
-  implicit val morphirIRVersionEncoder: JsonEncoder[MorphirIRVersion] = JsonEncoder.int.contramap(_.versionNumber.toInt)
+  implicit val morphirIRVersionEncoder: JsonEncoder[MorphirIRVersion] =
+    JsonEncoder.int.contramap(_.versionNumber.toDouble.toInt)
 
   implicit def morphirIRFileJsonEncoder: JsonEncoder[MorphirIRFile] =
     Json.encoder.contramap[MorphirIRFile] { file =>
